refactor: dedicated package for input formats
This allows to write a rule for import linter
which automatically extends on new packages
