+++ src/opossum_lib/input_formats/opossum/constants.py
