+++ src/opossum_lib/input_formats/opossum/opossum_file.py
