+++ src/opossum_lib/input_formats/opossum/opossum_file_content.py
@@ -11,9 +11,12 @@ from zipfile import ZipFile
 
 from pydantic import BaseModel, TypeAdapter
 
-from opossum_lib.opossum.constants import INPUT_JSON_NAME, OUTPUT_JSON_NAME
-from opossum_lib.opossum.opossum_file import OpossumInformation
-from opossum_lib.opossum.output_model import OpossumOutputFile
+from opossum_lib.input_formats.opossum.constants import (
+    INPUT_JSON_NAME,
+    OUTPUT_JSON_NAME,
+)
+from opossum_lib.input_formats.opossum.opossum_file import OpossumInformation
+from opossum_lib.input_formats.opossum.output_model import OpossumOutputFile
 
 
 class OpossumFileContent(BaseModel):
