+++ src/opossum_lib/input_formats/opossum/opossum_file_writer.py
@@ -4,12 +4,12 @@
 from pathlib import Path
 from zipfile import ZIP_DEFLATED, ZipFile
 
-from opossum_lib.opossum.constants import (
+from opossum_lib.input_formats.opossum.constants import (
     COMPRESSION_LEVEL,
     INPUT_JSON_NAME,
     OUTPUT_JSON_NAME,
 )
-from opossum_lib.opossum.opossum_file_content import OpossumFileContent
+from opossum_lib.input_formats.opossum.opossum_file_content import OpossumFileContent
 
 
 class OpossumFileWriter:
