+++ src/opossum_lib/input_formats/opossum/output_model.py
