+++ src/opossum_lib/input_formats/scancode/constants.py
