+++ src/opossum_lib/input_formats/scancode/model.py
