+++ tests/core/test_opossum_model.py
@@ -4,7 +4,7 @@
 import json
 from copy import deepcopy
 
-from opossum_lib.opossum.opossum_file_to_opossum_converter import (
+from opossum_lib.input_formats.opossum.opossum_file_to_opossum_converter import (
     OpossumFileToOpossumConverter,
 )
 from tests.test_setup.opossum_faker_setup import OpossumFaker
