+++ tests/test_opossum/generators/generate_file_information.py
@@ -13,7 +13,7 @@ from faker.providers.lorem.en_US import Provider as LoremProvider
 from faker.providers.misc import Provider as MiscProvider
 from faker.providers.person import Provider as PersonProvider
 
-from opossum_lib.opossum.opossum_file import (
+from opossum_lib.input_formats.opossum.opossum_file import (
     BaseUrlsForSources,
     ExternalAttributionSource,
     FrequentLicense,
