+++ tests/test_opossum/test_conversion_roundtrip.py
@@ -3,8 +3,8 @@
 # SPDX-License-Identifier: Apache-2.0
 from copy import deepcopy
 
-from opossum_lib.opossum.opossum_file_content import OpossumFileContent
-from opossum_lib.opossum.opossum_file_to_opossum_converter import (
+from opossum_lib.input_formats.opossum.opossum_file_content import OpossumFileContent
+from opossum_lib.input_formats.opossum.opossum_file_to_opossum_converter import (
     OpossumFileToOpossumConverter,
 )
 from tests.test_setup.opossum_file_faker_setup import OpossumFileFaker
