+++ tests/test_opossum/test_file_generation.py
@@ -5,9 +5,12 @@
 from pathlib import Path
 from zipfile import ZipFile
 
-from opossum_lib.opossum.constants import INPUT_JSON_NAME, OUTPUT_JSON_NAME
-from opossum_lib.opossum.opossum_file_content import OpossumFileContent
-from opossum_lib.opossum.opossum_file_writer import OpossumFileWriter
+from opossum_lib.input_formats.opossum.constants import (
+    INPUT_JSON_NAME,
+    OUTPUT_JSON_NAME,
+)
+from opossum_lib.input_formats.opossum.opossum_file_content import OpossumFileContent
+from opossum_lib.input_formats.opossum.opossum_file_writer import OpossumFileWriter
 from tests.test_setup.opossum_file_faker_setup import OpossumFileFaker
