+++ tests/test_scancode/model_helpers.py
@@ -5,7 +5,7 @@
 
 from pathlib import PurePath
 
-from opossum_lib.scancode.model import (
+from opossum_lib.input_formats.scancode.model import (
     Copyright,
     File,
     FileBasedLicenseDetection,
